test files

+++ 2-selection_sort.c
@@ -10,7 +10,7 @@
 
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min;
+	size_t i, j, min;
 
 	if (array == NULL || size == 0)
 		return;
